+++ src/main/kotlin/de/syncwork/hrworks/endpoints/absences/LeaveAccountData.kt
@@ -18,9 +18,9 @@ data class LeaveAccountDataRq @JvmOverloads constructor(
 
 @Serializable
 data class LeaveAccountData(
-    val holidayEntitlement: Int,
-    val requested: Int,
-    val approved: Int,
-    val unplanned: Int,
-    val planned: Int,
+    val holidayEntitlement: Double,
+    val requested: Double,
+    val approved: Double,
+    val unplanned: Double,
+    val planned: Double,
 )
